+++ src/main/java/com/bekvon/bukkit/cmiLib/CMIMaterial.java
@@ -1650,7 +1650,7 @@ public enum CMIMaterial {
             stack.setAmount(amount);
             return stack;
         }
-        ItemStack stack = new ItemStack(mat == null ? Material.STONE : mat, 1, (short) this.getLegacyData());
+        ItemStack stack = new ItemStack(mat == null ? Material.STONE : mat, 1, this.getLegacyData());
         stack.setAmount(amount);
         return stack;
     }
@@ -1671,7 +1671,7 @@ public enum CMIMaterial {
             stack.setAmount(amount);
             return stack;
         }
-        ItemStack stack = new ItemStack(mat == null ? Material.STONE : mat, 1, (short) this.getLegacyData());
+        ItemStack stack = new ItemStack(mat == null ? Material.STONE : mat, 1, this.getLegacyData());
         stack.setAmount(amount);
         return new CMIItemStack(stack);
     }
